Move CurlHandle into a separate file
